ai-law-compliance/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.26.0
chromadb==0.5.0
pinecone-client==3.2.2
sentence-transformers==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.3.0
structlog==24.2.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
